[Pricing-B2B] migrate B2B pricing from XML feeds to Listings API (#68)
* update B2B prices using listings-api

---------

+++ use-cases/pricing-b2b/code/java/src/main/java/lambda/utils/B2B/PricingRuleB2B.java
@@ -16,7 +16,7 @@ public class PricingRuleB2B {
     @JsonProperty("offerType")
     public String offerType;
     @JsonProperty("quantityTier")
-    public float quantityTier;
+    public int quantityTier;
     @JsonProperty("priceChangeRule")
     public String priceChangeRule;
     @JsonProperty("priceChangeRuleAmount")
